docs: Moves App component to component directory
