+++ src/components/App.js
@@ -1,6 +1,6 @@
-import './App.css';
-import BooksForm from './containers/BooksForm';
-import BooksList from './containers/BooksList';
+import '../App.css';
+import BooksForm from '../containers/BooksForm';
+import BooksList from '../containers/BooksList';
 
 const App = () => (
   <div className="App">
